chore: move normalizeClass to its own file
