+++ packages/@lwc/shared/src/index.ts
@@ -16,6 +16,7 @@ export * from './html-attributes';
 export * from './html-escape';
 export * from './meta';
 export * from './namespaces';
+export * from './normalize-class';
 export * from './overridable-hooks';
 export * from './static-part-tokens';
 export * from './style';
